Save and restore state from disk
This commit adds the ability for an agent, when restarted, to restore its state
from checkpoints on disk. This ensures it re-connects as the same
ContainerInstance and can report the state of containers it's managing.

This commit also introduces some tweaks to engine state transitions and
progression.

Create our own scratch container as well.