Remove dependency on docker in auth
The docker functions relied on were internal and have since changed
upstream.
This changes it to use the Kubernetes keyring which is a pkg, not an
internal detail.

This also improves how images are pulled slightly by doing less
image/tag parsing here and letting the docker daemon do more.