Engine: Don't let a 'stoptask' block itself
This fixes a bug whereby a task could prevent itself from stopping under
certain circumstances.

Basically, when the task is waiting for other tasks to stop before it
starts (the sequence number stuff), it did not read events off its
'acsChannel'. Because of that, a state transition of 'move to stopped'
for that task, if it occured before the stops the task was waiting for,
would result in a write to that channel which was not being read, and
thus lock up the entire system.

Namely, you ahve an event like the following:

TaskA: Start          - SequenceNumber 1
TaskA: Pull (slow operation)
TaskA: Stop           - SequenceNumber 1
       ^ Begin processing stop; waits on pull to complete.
TaskB: Start          - SequenceNumber 2
       ^ Wait for TaskA to stop
TaskB: Stop           - SequenceNumber 3
       ^ Takes processTasks lock and tries to write to TaskB, but B is
         not listening and the lock is never released.
TaskA: Stops
       ^ Too late to save the engine; processTasks lock taken *forever*

The solution to this is to always read new events off the acsMessage
channel (it already read them in literally every other case) and to just
shortcut out if you change to terminal.

The other changes (Mutex -> RWMutex) don't matter here I think, but I
felt that there was a sufficient distinction there that it was okay to
have it.