Engine: Skip pulls when desired stopped.
When you recieve an event to 'stop' a task, the pulls no longer really
matter. This lets you break out in that case. It also includes allowing
some other shortcutting (e.g. don't try to stop containers that are
known to be not running) which has a nice side effect of toning down on
some 'reasons' being reported even during normal operation..