Config: Do not compute endpoint; make region fatal
The SDK can handle setting the endpoint better than the Config code did;
the endpoint behavior before was a legacy of a time with no sdk.

The specific issue was that the endpoint would be set incorrectly if the
ec2 metadata service did not return any value.

In line with this, if no region can be determined it now is a fatal
error.

Relates to #74