Fix bug with BindIP reporting
When iterating through a slice, the value variable gets reused for each
iteration.  While the values contained in that variable change with
iteration, the address does not.  Assigning the address of that variable
in the loop allows the actual value to change out from under, meaning
that the last iteration through the loop wins.