ACS: Handle heartbeats vs idle correctly
Previously a heartbeat message was required to consider the channel
active.
In realitiy, heartbeat messages were only sent when the channel was
inactive and no other messages were being sent.
This avoids treating a lack of heartbeats as an idle channel and closing
it unless there are also no other messages.

In addition, this tweaks how backoffs are handled (time, resets, etc) a
bit to be more forgiving to these sorts of issues (where the connection
is lost, but can be re-established).

Relates to #103