Reregistration: Don't ignore errors
Even if we've registered in the past and have a saved arn, avoid running
if we cannot do a re-register. This typically indicates a networking
error and we won't be able to do meaningful work in that state anyways.

This ensures better correctness with respect to registered resources and
so on as if they mutate, we won't recieve work until we report that
successfully.