Engine: Improve exitcode detection and logging
Prior to this change, the 'exitcode' would be set in any case where the
container was not running, including created. This scopes it down to
only be after it exists. This also improves logging.

In addition, this ensures we log the DockerID and name explicitly at
info level since recovering that mapping from logs is often valuable