Rework handler tests
This adds more mocking and makes them pass more reliably.

This does involve some type changes / casting as well to allow mocking,
but the changes are primarily testing related.

This removes the previous testing of actually spinning up an http server
and making requests and replaces it with lots of http recording of
handlers