Clean up the PR for task cleanup:
1. Rename config.CleanupWaitDuration to config.TaskCleanupWaitDuration to be more
explicit about its intent.
2. Add a minimum threshold for the cleanup time. Accepting low values
such as '1 us' causes the agent to deadlock.
3. Fix a bug in the engine code which could cause a deadlock
when using en empty config.Config object in the engine. The waitEvent
method will deadlock when waiting for cleanup if the cleanup go routine
exits prematurely. A check has been added to the event duation to fix
this.
4. Add functional tests for task cleanup.