engine: Disable OOM test for broken kernel configs
This test is only reliable when the kernel properly supports memory and
swap limits.  If you try to limit memory on a system where the kernel
does not support swap limits, the container can continue to use memory
in swap well beyond the expected timeout of the test.  If your system
does not support swap limits, you'll see a message like the following
when running a container with a memory limit:

WARNING: Your kernel does not support swap limit capabilities, memory limited without swap.

Note that you should not disable this test by default.