[BugFix] Reset ACSSeqNum stored in state file when containerInstance is reset
This fix will prevent from agents connecting using bad ACSSeqNum when agent is registered as new Container Instance.
This usually happens when EC2 instance (registered as Container Instance) is snapshotted and launched as a new EC2 instance i
(registered as new Container Instance). Before this fix the ACSSeqNum of the previous Container Instance is used for the new one.