Enable Task IAM Role for containers launched with 'host' network mode. (#52)
* Add OUTPUT rule for credentials endpoint.

Modify netfilter Create and Remove functions to add and remove
an OUTPUT rule to redirect processes/containers that use
host's network stack to redirect credentials requests to
localhost:51679.

This is needed for ensuring that containers launched with
net=host can still talk to the credentials endpoint.

* Set ECS_ENABLE_TASK_IAM_ROLE_NETWORK_HOST=true for ECS Agent.

Set the ECS_ENABLE_TASK_IAM_ROLE_NETWORK_HOST env var config to true
when starting the ECS Agent. This config registers the ECS Agent with
the capability to run Tasks that require Task IAM Roles when
created with 'host' network mode.