Enable Task IAM Role for containers launched with 'host' network mode. (#468)
* Enable Task IAM Role for containers launched with 'host' networkmode.

Register a new capability for the Container Instance to indicate if
it is capable of launching Tasks that need an IAM Role as well as
host network mode.

Create an env var config for the same as well

* Remove force tagging images when setting up test registry.

Following the deprecation notice for docker tag -f, remove the
flag in the script. Refer https://docs.docker.com/engine/deprecated/#/f-flag-on-docker-tag
for details.