engine: Set inactivity timeout for stats stream
Our vendored version of go-dockerclient does not actually close the
stream when we signal that we're done reading it, leading to a potential
connection leak over time.  Setting an inactivity timeout will allow
this version of go-dockerclient to disconnect once activity ceases
across the stats stream (such as when the container stops).