Bugfix: Stats engine reconciliation for missed remove events.
This is a quick fix for cases where stats engine might miss
'container removed' event from the docker task engine
event stream.

Without this fix, the stats engine will spin on
'docker stats' api until the container is removed.

This is mentioned as a quick fix because this is not addressing the
root-cause of why an event is sometimes being missed by the
stats engine.

There's also the case of 'container addeded' event being missed
which could cause metrics to be misrepresented.

This is also proposed as a fix for the issue #515.

There's a minor refactor of the task engine code where locks have been
added to protect api.Container status'es