export port used for IAM credentials
The documentation for using IAM roles with docker tasks
states that when you are not using the amazon provided ECS OPTIMIZED
image, you need to fiddle with the iptables to route the special
IP 169.254.170.2 PORT 80 to 127.0.0.1:51679

https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task-iam-roles.html

However, this port is not listed as an exposable port via docker inspect
so it took me a sec to figure it out.  You can map the port manually anyway,
but this change makes things a little more obvious.

Now I'm not sure if this port isn't actually started in the event IAM
roles are not used, but this change shouldn't affect that since nothing
will be listening and the difference is zip.