Fix event state race condition in task engine unit tests.
I think this originated from copying test code from a unit test
that already had this bug because of the race condition. The deleted
line of code used to result in race condition and it would trigger
the task engine to do additional work after the test would complete
leading to test failures.

Should fix #669