engine: Report task status prior to removal
Prior to this change, a race condition exists between reporting task
status and task cleanup in the agent.  If reporting task status took an
excessively long time and ECS_ENGINE_TASK_CLEANUP_WAIT_DURATION is set
very short, the containers and associated task metadata could be removed
before the ECS backend is informed that the task has stopped.  When the
task is especially short-lived, it's also possible that the cleanup could
occur before the ECS backend is even informed that the task is running.
In this situation, it's possible for the ECS backend to re-transmit the
task to the agent (assuming that it hadn't started yet) and the
de-duplication logic in the agent can break.

With this change, we ensure that the task status is reported prior to
cleanup actually starting.