engine: fix regression in TestRemoveEvent test
This test regressed with PR #706. The original intent of the test was to test
if  gets exercised while a task is being cleaned up. This
was accomplished by sending events to the engine's container event stream while
cleanup was in progress using the mock client's  method.

However, we lost this part with PR#706. This commit adds that code back and
removes some of the unnecessary code in .