api, statemanager: refactor 'api.Container.IsInternal' to an enum
The IsInternal field in api.Container is being refactored as an enum instead
(as 'api.Container.Type ContainerType'), with values that indicate what type of
container it is. It can be one of 'ContainerNormal' (indicates that this is
not an internal container, but something that was sent as part of the payload
from the communication service) or 'ContainerEmptyHostVolume' (indicates that
this is an internal container created for attaching ephemeral empty host
volumes) or 'ContainerCNIPause' (indicates that this is an internal container
created for attaching ENIs or some other custom network configuration)

The 'api/json.go' file has been deleted and its contents moved to files for
each respective type that has custom marshal/unmarshal methods overridden.

The version of the ECS Data file has been incremented as well.