api, eventhandler: updated models, added batching
This change updates the model for the ecs client and includes the
corresponding code changes to SubmitTaskStateChange. Secondly we've
added the batching logic to eventhandler, which collects container state
change events as the generated and then attaches the collection of
events to the SubmitTaskStateChange payload as dictated but the model
changes.