agent/app: wait for child processes and SIGCHLD
Since dhclient processes for ENIs are launched as child processes
of the Agent process, they will be zombied if Agent does not
wait for its child processes to exit. This also means that we may
have to write a custom cmd.Wait wrapper where we ignore
os.SyscallError for the 'wait' syscall.