agent/app: get rid of the reaper.
The CNI plugin invocation by the ECS Agent fails if the Agent gets a `SIGCHLD`
while the CNI plugin invocation is in progress because of the overarching
`waitpid(-1)` reaper code. In this scenario, the catch-call `waitpid(-1)` causes
CNI's `ExecPlugin` call to return with `waitid` syscall error because it
relies on [cmd.Run()](https://github.com/containernetworking/cni/blob/master/pkg/invoke/raw_exec.go#L42),
which in turn calls [cmd.Wait](https://golang.org/pkg/os/exec/#Cmd.Wait),
leading to the failure described above.

It could be argued that Agent can be smarter and wait only for `dhclient`
processes by starting them in a different process group and hence not
needing to do `waitpid(-1)`. However, I'm not sure if it's feasible for a child
process to  be created with a process group that's different from that of the parent
(Agent). An `init` system would still be needed to accomplish that.

It is much cleaner to to instead run the ECS Agent with host's
pid namespace. In the future, we can revisit this decision if we want to
run the ECS Agent with an `init` system. The host's `/proc` directory will still
mounted as `/host/proc` within the Agent container as per
https://github.com/aws/amazon-ecs-agent/blob/enis/agent/ecscni/types.go#L32
and that's how it accesses the `netns` handle for the `pause` container.
The only change in the future would be to get rid of the `--pid=host`
when an Agent is configured to run with an init system. This also gives us
the flexibility of not gating the "Task ENIs" feature to only those versions
of docker that support running an init system and doesn't bloat the Agent
itself with an `init` system.