Fix type assertion bug that could lead to reporting of incorrect task status.
A type assertion of event.Error assumed that the variable being tested was a
reference, when we actually weren't consistent in how we assigned that
variable. This caused the assertion to fail in places where it should have
passed, ultimately causing us to falsely report containers as stopped in the
case where docker returned an API error in response to the 'stop' call.

To fix this, we define a new 'cannotStopContainerError' interface and perform
the type assertion on the interface.