Do not explicitly set host IP to 0.0.0.0 for Linux, defer to Docker
---* Motivation *---
By hardcoding the host IP as 0.0.0.0, the agent is removing the ability for customers to specify specific host IP addresses (such as when there are multiple network interfaces). By default, the docker daemon defaults to 0.0.0.0, but this value may be overridden by launching the daemon with a --ip flag. https://docs.docker.com/engine/reference/commandline/dockerd/

This also has an implication for the CloudWatch Events published by ECS. Specifically, prior to this change the 'bindIP' field in the ECS Task Status events always contains '0.0.0.0'.

This change will benefit the following customer use cases:
* https://github.com/aws/amazon-ecs-agent/issues/426
* https://forums.aws.amazon.com/thread.jspa?threadID=173250
* https://serverfault.com/questions/695393/amazon-ecs-docker-binding-container-to-specific-ip-address

---* Detailed Description *---
Do not explicitly set the HostIP property when starting containers and instead defer to the daemon's default.