Support reserving CPU units for the system
This adds support for reserving CPU units for the system, very similar to the reserved memory. ECS_RESERVED_CPU is a new environment var that accepts CPU units, which are subtracted from the hosts total resource amount. This patch also switches the internals of ECS_RESERVED_MEMORY to use a 32-bit integer (previously 16-bit) so that it can support a reservation of more than 64GB.