engine: "internal" containers may not be pulled
This commit treats internal containers differently during the pull step
of the container lifecycle.

We currently have two types of internal containers: "empty host volume"
containers and pause containers.  "Empty host volume" containers should
either be pulled from a remote image (as on Windows) or imported from a
locally-generated tarball (as on Linux).  The pause container image is
either imported from a locally-embedded tarball inside the agent's
rootfs or expected to already be present on the host (if the image is
overridden via config).