eventhandler: fix events batching race
There was another race condition where, if an event was already
being marked for submission, the batched container changes would
be lost if another submission trigger came in for the same task
from the ticker. This commit makes it so that the two maps
in task handler are updated in a single lock scope to avoid such
a race.

This commit also refactors the code to get rid of most methods
that were not tied to any struct. Instead, each struct
has its own methods to deal with sending state changes to ECS