Fix timing issue with hostsetup script
I noticed that sometimes the port forwarding that we do using netsh doesn't work. The ECS agent does listen on port 51679, but we cannot connect to it using port 80, and hence IAM task roles don't work. I figured out that this happens when the netsh command is executed before the Credential IP is up. Maybe this is because we do the IP allocation in powershell, and then use netsh (cmd) to do the power forwarding. This change makes sure that the IP is up and responding before we try to set up port forwarding. And with this, I haven't noticed the issue anymore.