Remove task/container overridden in DockerHostConfig
In task/container overridden, it's copying the task/container struct
which has a lock, this could cause deadlock if some other go routine
acquired the lock right before copying and released the lock after the
copy. This commit moved the overridden part to ACS when unmarshal the
task from payload message.