stats(engine): refactor a bit to reorder locks
locking in this file was a bit erratic. This refactor hopefully
makes things easier to understand. methods that access shared data
without a lock are renamed with "Unsafe" suffix. The taskToContainers
map was being accessed in a way that could result in a data race.
This is fixed as well.