Expose task definition family and revision as metadata fields
The task struct already has this information, so it's just a matter
of plumbing it through. Note that the type of the revision field in
the struct is a string, and the type in our CLI is an integer
(https://docs.aws.amazon.com/cli/latest/reference/ecs/describe-task-definition.html).
This change is consistent with the CLI.

(cherry picked from commit fdf1810325023f9f64e63eb0462c5266d4d056c4)
Signed-off-by: Josh Oberwetter <oberj@amazon.com>