Systemd cgroup support
When running docker under systemd docker requires that --cgroup-parent
matches the systemd hierarchy. Otherwise it docker run will fail with
the message:

cgroup-parent for systemd cgroup should be a valid slice named as
"xxx.slice"

This creates the root cgroup ecs.slice and runs each task in a subslice
ecs-<task-id>.slice (with - replaced with _).

This also adds ECS_CGROUP_PREFIX to change the name of the root slice.