refactor task engine unit tests
This commit moves commonly used code in the task engine unit tests
into methods. It also reuses these methods in various places to
increase the readability of the tests. It also deletes a number of
redundant tests and adds a new table based test for the happy path
test.

Test covered has now increased to 79.3% from 77.8% for this package