engine: make the container progress concurrently
Previously the waitForContainerTransition will wait for all the
transitions to be done before the container can move to next state. This
PR changes the waitForContainerTransition to only wait for no more than
one transition to be done so that the first container that done the
transition doesn't need to wait for other containers.