dockerclient: add timeout for Version API
This change adds a timeout for Docker client's Version API.
Additionally, the Docker version returned in saved in the task engine.
This helps with avoiding additional unnecessarily calls to the Docker
Version API everytime the agent connects to the backend. This should
help with issues seen in #1331

Finally, the dockerclient package has been refactored to extract all
timeout related constants to dockerclient package from the dockerapi
package to avoid circular dependencies.