config: add NoIID config variable
`NoIID` when set specifies that ECS agent should not try to retrieve
instance identity document in the 'register' path. This is different
than the '-k' agent start-up option, which specifies that the instance
metadata endpoint should not be accessed at all. This is useful in cases
where we want ECS agent registration to succeed when running on
infrastructure that does not support retrieving instance identity
documents.