Changing ecr image pull retry to 5
Reducing the number of retry calls made by agent to pull a docker image.

Docker has it's own getLayer retry in the daemon. Including that in the worst case, agent will make 10 * 5 * nLayers calls(nLayers is the number of layers in the image). This change will bring down the # of calls to 5 * 5 * nLayers.