Remove the functionality of StartTimeout as Docker API Start Timeout
The ‘StartTimeout’ now will only serve as the the time duration after which if a container has a dependency on another container and the conditions are ‘SUCCESS’, ‘HEALTHY’ and ‘COMPLETE’, then the dependency will not be resolved. For example:

•	If a container A has a dependency on container B and the condition is ‘START’, the StartTimeout for container B will roughly be the time required for it to exit successfully with exit code 0
•	If a container A has a dependency on container B and the condition is ‘COMPLETE’, the StartTimeout for container B will roughly be the time required for it to exit.
•	If a container A has a dependency on container B and the condition is ‘HEALTHY’, the StartTimeout for container B will roughly be the time required for it to emit a ‘HEALTHY’ status.

If the StartTimeout exceeds in any of the above cases, container A will not be able to transition to ‘CREATED’ status.

It effectively reverts the implementation of StartTimeout in commit: 79bd51770754121c4be28328a0f683c2526e18c8