tests: make windows test run a bit faster
* Remove need to pull 'latest' server core

By removing the :latest tag from all windowsservercore containers,
we will have the tests use the container thats already baked into
the AMI.

* Remove depdency on golang and python containers

We are removing the need to use any containers other than servercore
and nanoserver. This reduces the number of downloads needed and the
number of builds that happen before the tests start running.

* Explicit timeouts on order tests

The ordering tests are broken at the moment, so we are capping them
with a fixed timeout.