Fix TestOOMContainer on Windows.
TestOOMContainer runs a task that's expected to fail fast because the contain tries to use too much memory. The test had a check where it was waiting for the task to reach running, but since the task stops quickly, it's possible that when the test describes the task, the task has already stopped and the test fails. Fixing by removing the wait running check.