Fix TestDockerStateToContainerState integ test.
During the workflow of TestDockerStateToContainerState, it starts a container and checks whether it's running by inspecting it. However, the container it starts exit immediately, so it's possible that when we check whether the container is running it's already stopped. This leads to flakey test as per https://github.com/aws/amazon-ecs-agent/issues/1899. Fixing by letting the container keep running rather than exit itself.