functional_tests: remove dependency on python
This removes the need to use python:2. The only test we required python for
boiled down to 'allocate X bytes'. The telemetry test already does that exact
thing.

In order to make this a bit more clear, I've renamed the 'telemetry-test'
container to 'amazon-ecs-stress', which reflects the name of the binary it
uses.