Wait status active/inactive when task networking is enabled.
When task networking is enabled and instance type supports eni trunking, eni trunking will be turned on. In that case, there will be a "REGISTERING" status for container instance where we attach the trunk ENI and no task can be placed on it during that time, and a "DEREGISTERING" status where the trunk is detached. We need to wait for container instance to reach ACTIVE to avoid start task failure, and wait for container instance to reach INACTIVE to ensure the trunk is detached.