Refactor task ENI setup
ACS payload message can send a list of network interface devices and
a list of ipv4 addresses per network interface to the agent. The
current implementation considers a message with where there's a
plurality of interfaces and ipv4 addresses to be invalid messages. This
might not always be the case. This commit introduces modifications to
this logic where such messages are gracefully handled.

This commit also refactors the `ecscni` package, where all of the config
generation logic is moved out of `plugin.go` into `netconfig.go`. This
ensures that all of the logic that deals with various CNI plugins gets
centralized in the same place instead of being spread out over different
files.