efs: adding support for additional nfs options
NFSMount now allows for "ReadOnly" and "AdditionalOptions" to be set on the
struct. If "AdditionalOptions" is set, agent will combine the user-provided
options with its defaults. For now, we only allow NFSv4 options, and not
general mount options. Options are a subset of the options listed in nfs(5).

ReadOnly will mount the volume with the RDONLY flag if specified.