api/task: change type of ENI field
Modified the type of `ENI` field in `task.Task` from eni.ENI to
task.TaskENIs. The new type is an alias for []eni.ENI. It's helpful
when dealing with state files from existing agents, where the ENI
field is currently being stored as a scalar. The TaskENIs type's
Unmarshal() method is capable of decoding both scalar and list types.

Also, fixed bugs with the previous commit where some of the cni confg
fields were not being properly populated.