support JSON key and version for secretsmanager secrets specified
for roadmap issue #385
https://github.com/aws/containers-roadmap/issues/385
this commit adds the ability for customers to add parameters
to the secretsmanager ARN specified in containers. agent will be
able to retrieve secret by version or retrieve part of a secret
by json key.
this commit also fixes a minor issue breaking go vet in an unrelated test.