Don't clear stats queue every interval
Implications of this change:
1. During one TACS publishing interval, we previously would only send
metrics to TACS if we received two separate stats from docker. Going
forward, if we only received a single stat we will compare it to a stat
from a previous interval in order to get the usage.
2. This also allows the user to configure the stats collection interval
to be once per TACS publish interval, further improving CPU usage.

Fix unit tests