Explicitly initialize seelog
The logger package's initialization logic is init() which will be
executed by loading the package.

However some of our programs are using the agent as a library and
haven't configured seelog beforehand. Due to that, loading the package
logs

```
1600397429848419666 [Error] node must have children
```
which breaks the programs.

I think it would be safer to do less on init() and explicitly
initialize the logger from the agent's main().

Signed-off-by: Kazuyoshi Kato <katokazu@amazon.com>