Add and use new instancecreds package
This is to centralize grabbing our instance credential provider into a
single package. Previous to this change we grabbed instance credentials
in separate places for the ecs, ec2, ecr and ec2 metadata clients. With this
change, all of these clients will use the same instance credential
provider.

The reason for this is that if we want to add any non-default instance
credential providers to the credential chain, then we need to make sure
that all AWS clients are instantiated with the same provider.