scripts: add ci-ecr-pull and ci-ecr-push
These scripts are for pushing the dockerhub images that we need for our
ci system (integ test and build) to ECR, and for pulling those images
down within the CI system.

ci-ecr-pull will tag the ECR images like 'golang:1.15.4' so that users
who do not run this script on their laptop or elsewhere can still pull
from dockerhub, only if this script is run will the builds reference the
ECR images.

The reason behind this is that dockerhub is now rate-limiting anonymous
pulls, so in order to avoid being rate-limited we should pull our images
from ECR as much as we can.

This should also speed up builds because we will be pulling our large
build images (golang, gcc) from in-region s3 rather than over the
internet.

One side-effect of this change is that we have to reference full version
numbers in our dockerfiles (1.15.4 instead of 1.15), in order to avoid
querying docker for the latest version, and to avoid dockerhub pulls
happening when an image is updated.