Use SSM vault file to get managed instance id
The file at /var/lib/amazon/ssm/registration is created by the ssm agent
but is not subsequently used. As such, when telling the ssm agent to
clear registration using 'amazon-ssm-agent -register -clear', this file
is not cleared.

The SSM vault file is actively managed by the SSM agent and it is
cleared when the instance is deregistered.