added support for agent to advertise os-family attribute to the control plane
In order to allow fine grain task placement, a new attribute "os-family" is being added which will allow the customers to place their task on specific instances running in a mixed mode cluster.

For Windows, the value of this attribute would be similar to WINDOWS_SERVER_2019_FULL. The same can be specified by the customer in their task definition. Now, if a cluster runs instances running on different Windows versions, this task would only be placed on the instances which broadcasted their os-family as WINDOWS_SERVER_2019_FULL.