DiscoverPollEndpoint: lengthen cache ttl and improve resiliency (#3109)
The acs/tacs endpoints that ecs agent connects have never changed
before, so the current behavior of calling the API every 20 minutes is
not necessary. Lengthening this cache TTL to 12 hours will reduce load
on the ECS service and reduce the chance of customers being throttled.

Additionally, the LRU cache we were using automatically evicted any
value once it expired. This means that this API call could become a
source of complete ecs agent failure in the event of an LSE, in the
event that agent threw away it's cached endpoint and then failed to get
a new one via DiscoverPollEndpoint.

By keeping expired values in the cache, we can fallback to use the
expired value in the event that the DiscoverPollEndpoint API is failing,
thus keeping agent connected to ACS and TACS.